pyproject.toml
[project]
name = "sprite-me"
version = "0.1.0"
description = "Agent-first pixel art sprite generator via MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "rembg>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
]

[project.scripts]
sprite-me-server = "sprite_me.server:main"
sprite-me-api = "sprite_me.api:main"
sprite-me-volumes = "sprite_me.volumes:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sprite_me"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
